src/test/java/com/mygaienko/common/algorithms/leetcode/word_search_ii/SolutionTest.java
package com.mygaienko.common.algorithms.leetcode.word_search_ii;

import org.junit.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;


public class SolutionTest {

    Solution solution = new Solution();

    @Test
    public void test() {
        List<String> words = solution.findWords(new char[][]{
                {'o', 'a', 'a', 'n'},
                {'e', 't', 'a', 'e'},
                {'i', 'h', 'k', 'r'},
                {'i', 'f', 'l', 'v'}
        }, new String[]{"oath", "pea", "eat", "rain"});
        assertThat(words).containsExactlyInAnyOrder("eat", "oath");
    }
}